Renaming HierarchialLogger packages
...so that it is consisten with arena packages
